@@ -1,16 +1,16 @@
 ## Opscode Expander ##
-source "http://rubygems.org"
+source 'http://rubygems.org'
 
-gem "amqp", "~> 0.6.7"
-gem "eventmachine", '~> 1.2.0'
-gem "em-http-request", "~> 0.2.11"
-gem 'yajl-ruby', "~> 0.7.7"
-gem 'mixlib-log', "~> 1.1.0"
-gem 'uuidtools', "~> 2.1.1"
+gem 'amqp', '~> 0.6.7'
+gem 'eventmachine', '~> 1.2.0'
+gem 'em-http-request', '~> 0.2.11'
+gem 'yajl-ruby', '~> 1.3.0'
+gem 'mixlib-log', '~> 1.1.0'
+gem 'uuidtools', '~> 2.1.1'
 gem 'bunny', '~> 0.6.0'
-gem 'fast_xs', "~> 0.7.3"
+gem 'fast_xs', '~> 0.7.3'
 gem 'highline', '~> 1.6.1'
-gem 'rake', '~> 0.8.7'
-gem 'rspec', '~> 1.3.0'
+gem 'rake', '>= 0.8', '< 13.0'
+gem 'rspec', '~> 1.3'
 gem 'word-salad', '~> 1.0.0'
 gem 'veil', git: 'https://github.com/chef/chef_secrets'
