@@ -6,7 +6,7 @@ override :lua, version: "5.1.5"
 override :'omnibus-ctl', version: "master"
 override :chef, version: "v12.19.36"
 override :ohai, version: "v8.23.0"
-override :ruby, version: "2.2.6"
+override :ruby, version: "2.4.1"
 override :rubygems, version: "2.6.8"
 # This SHA is the last commit before the 6.0 release
 override :'berkshelf-no-depselector', version: '6016ca10b2f46508b1b107264228668776f505d9'
